+++ src/vfvm_linsolve.jl
@@ -4,28 +4,33 @@ canonical_matrix(A) = A
 canonical_matrix(A::AbstractExtendableSparseMatrixCSC) = SparseMatrixCSC(A)
 
 function _solve_linear!(u, state, nlhistory, control, method_linear, A, b, reuse_precs)
+
     if isnothing(state.linear_cache)
         if !isa(method_linear, LinearSolve.SciMLLinearSolveAlgorithm)
             @warn "use of $(typeof(method_linear)) is deprecated, use an algorithm from LinearSolve"
         end
-        Pl = nothing
         nlhistory.nlu += 1
-        p = LinearProblem(canonical_matrix(A), b)
-        state.linear_cache = init(
-            p,
-            method_linear;
-            abstol = control.abstol_linear,
-            reltol = control.reltol_linear,
-            maxiters = control.maxiters_linear,
-            verbose = doprint(control, 'l'),
-            Pl,
-        )
+        nlhistory.tlinsolve_setup += @elapsed begin
+            Pl = nothing
+            p = LinearProblem(canonical_matrix(A), b)
+            state.linear_cache = init(
+                p,
+                method_linear;
+                abstol = control.abstol_linear,
+                reltol = control.reltol_linear,
+                maxiters = control.maxiters_linear,
+                verbose = doprint(control, 'l'),
+                Pl,
+            )
+        end
         if doprint(control, 'l')
             out = @sprintf("    [l]inear: factorize #%d\n", nlhistory.nlu)
             _info(out)
         end
     else
-        reinit!(state.linear_cache; A = canonical_matrix(A), b, reuse_precs)
+        nlhistory.tlinsolve_setup += @elapsed begin
+            reinit!(state.linear_cache; A = canonical_matrix(A), b, reuse_precs)
+        end
         if !reuse_precs
             nlhistory.nlu += 1
             if doprint(control, 'l')
@@ -36,9 +41,11 @@ function _solve_linear!(u, state, nlhistory, control, method_linear, A, b, reuse
     end
 
     try
-        sol = LinearSolve.solve!(state.linear_cache)
+        local sol
+        nlhistory.tlinsolve_solve += @elapsed begin
+            sol = LinearSolve.solve!(state.linear_cache)
+        end
         u .= sol.u
-        nliniter = sol.iters
         nlhistory.nlin = sol.iters
     catch err
         if (control.handle_exceptions)